Adding 3 modules + fixing CROS SUPPORT

+++ Core/Entities/Product.cs
@@ -6,7 +6,7 @@
 
         public string Description { get; set; }
 
-        public decimal ? Price { get; set; }
+        public float ? Price { get; set; }
 
         public string PictureUrl { get; set; }
